# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AISummary": {
        "properties": {
          "architecture": {
            "maxLength": 2500,
            "title": "Architecture",
            "type": "string"
          },
          "commit_sha": {
            "title": "Commit Sha",
            "type": "string"
          },
          "disclaimer": {
            "title": "Disclaimer",
            "type": "string"
          },
          "entry_points": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Entry Points",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "key_modules": {
            "items": {
              "$ref": "#/components/schemas/KeyModule"
            },
            "maxItems": 10,
            "title": "Key Modules",
            "type": "array"
          },
          "maintenance_concerns": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "title": "Maintenance Concerns",
            "type": "array"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "onboarding_steps": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "title": "Onboarding Steps",
            "type": "array"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "purpose": {
            "maxLength": 1500,
            "title": "Purpose",
            "type": "string"
          }
        },
        "required": [
          "purpose",
          "architecture",
          "key_modules",
          "entry_points",
          "maintenance_concerns",
          "onboarding_steps",
          "provider",
          "model",
          "commit_sha",
          "generated_at",
          "disclaimer"
        ],
        "title": "AISummary",
        "type": "object"
      },
      "AnalysisLimits": {
        "properties": {
          "max_file_bytes": {
            "title": "Max File Bytes",
            "type": "integer"
          },
          "max_files_to_fetch": {
            "title": "Max Files To Fetch",
            "type": "integer"
          },
          "max_total_fetch_bytes": {
            "title": "Max Total Fetch Bytes",
            "type": "integer"
          },
          "max_tree_entries": {
            "title": "Max Tree Entries",
            "type": "integer"
          }
        },
        "required": [
          "max_tree_entries",
          "max_files_to_fetch",
          "max_file_bytes",
          "max_total_fetch_bytes"
        ],
        "title": "AnalysisLimits",
        "type": "object"
      },
      "AnalysisMeta": {
        "properties": {
          "analyzed_at": {
            "format": "date-time",
            "title": "Analyzed At",
            "type": "string"
          },
          "commit_sha": {
            "title": "Commit Sha",
            "type": "string"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "engine_version": {
            "title": "Engine Version",
            "type": "string"
          },
          "fetch": {
            "$ref": "#/components/schemas/FetchSummary"
          },
          "github_api_calls": {
            "title": "Github Api Calls",
            "type": "integer"
          },
          "head_commit_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Commit Date"
          },
          "limits": {
            "$ref": "#/components/schemas/AnalysisLimits"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "partial": {
            "title": "Partial",
            "type": "boolean"
          },
          "rate_limit": {
            "$ref": "#/components/schemas/RateLimitInfo"
          }
        },
        "required": [
          "analyzed_at",
          "duration_ms",
          "commit_sha",
          "head_commit_date",
          "github_api_calls",
          "rate_limit",
          "fetch",
          "limits",
          "partial",
          "notes",
          "engine_version"
        ],
        "title": "AnalysisMeta",
        "type": "object"
      },
      "AnalysisReport": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/AnalysisMeta"
          },
          "architecture": {
            "$ref": "#/components/schemas/ArchitectureAnalysis"
          },
          "dependencies": {
            "$ref": "#/components/schemas/DependencyAnalysis"
          },
          "documentation": {
            "$ref": "#/components/schemas/DocumentationAnalysis"
          },
          "health": {
            "$ref": "#/components/schemas/HealthReport"
          },
          "languages": {
            "$ref": "#/components/schemas/LanguageAnalysis"
          },
          "quality": {
            "$ref": "#/components/schemas/QualityAnalysis"
          },
          "repository": {
            "$ref": "#/components/schemas/RepositoryMetadata"
          },
          "schema_version": {
            "default": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "security": {
            "$ref": "#/components/schemas/SecurityAnalysis"
          },
          "structure": {
            "$ref": "#/components/schemas/StructureAnalysis"
          }
        },
        "required": [
          "repository",
          "analysis",
          "health",
          "structure",
          "languages",
          "dependencies",
          "quality",
          "documentation",
          "security",
          "architecture"
        ],
        "title": "AnalysisReport",
        "type": "object"
      },
      "AnalyzeRequest": {
        "properties": {
          "repository_url": {
            "examples": [
              "https://github.com/psf/requests"
            ],
            "maxLength": 512,
            "minLength": 1,
            "title": "Repository Url",
            "type": "string"
          }
        },
        "required": [
          "repository_url"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
      "ArchitectureAnalysis": {
        "properties": {
          "compose_services": {
            "items": {
              "$ref": "#/components/schemas/ComposeService"
            },
            "title": "Compose Services",
            "type": "array"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/ModuleEdge"
            },
            "title": "Edges",
            "type": "array"
          },
          "entry_points": {
            "items": {
              "$ref": "#/components/schemas/EntryPoint"
            },
            "title": "Entry Points",
            "type": "array"
          },
          "frameworks": {
            "items": {
              "$ref": "#/components/schemas/Detection"
            },
            "title": "Frameworks",
            "type": "array"
          },
          "hubs": {
            "items": {
              "$ref": "#/components/schemas/FileHub"
            },
            "title": "Hubs",
            "type": "array"
          },
          "import_resolution": {
            "$ref": "#/components/schemas/ImportResolution"
          },
          "infrastructure": {
            "items": {
              "$ref": "#/components/schemas/InfraItem"
            },
            "title": "Infrastructure",
            "type": "array"
          },
          "methodology": {
            "title": "Methodology",
            "type": "string"
          },
          "modules": {
            "items": {
              "$ref": "#/components/schemas/Module"
            },
            "title": "Modules",
            "type": "array"
          },
          "monorepo_tool": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monorepo Tool"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "project_types": {
            "items": {
              "$ref": "#/components/schemas/Detection"
            },
            "title": "Project Types",
            "type": "array"
          },
          "workspace_packages": {
            "items": {
              "type": "string"
            },
            "title": "Workspace Packages",
            "type": "array"
          }
        },
        "required": [
          "project_types",
          "frameworks",
          "monorepo_tool",
          "workspace_packages",
          "entry_points",
          "modules",
          "edges",
          "hubs",
          "compose_services",
          "infrastructure",
          "import_resolution",
          "methodology",
          "notes"
        ],
        "title": "ArchitectureAnalysis",
        "type": "object"
      },
...
      "DependencyAnalysis": {
        "properties": {
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/Dependency"
            },
            "title": "Dependencies",
            "type": "array"
          },
          "dependencies_truncated": {
            "title": "Dependencies Truncated",
            "type": "boolean"
          },
          "development": {
            "title": "Development",
            "type": "integer"
          },
          "ecosystems": {
            "items": {
              "$ref": "#/components/schemas/EcosystemSummary"
            },
            "title": "Ecosystems",
            "type": "array"
          },
          "external_sources": {
            "title": "External Sources",
            "type": "integer"
          },
          "lockfiles": {
            "items": {
              "type": "string"
            },
            "title": "Lockfiles",
            "type": "array"
          },
          "manifests": {
            "items": {
              "$ref": "#/components/schemas/ManifestSummary"
            },
            "title": "Manifests",
            "type": "array"
          },
          "manifests_not_analyzed": {
            "items": {
              "type": "string"
            },
            "title": "Manifests Not Analyzed",
            "type": "array"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "other": {
            "title": "Other",
            "type": "integer"
          },
          "production": {
            "title": "Production",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "unconstrained": {
            "title": "Unconstrained",
            "type": "integer"
          },
          "unique_packages": {
            "title": "Unique Packages",
            "type": "integer"
          },
          "version_conflicts": {
            "items": {
              "$ref": "#/components/schemas/VersionConflict"
            },
            "title": "Version Conflicts",
            "type": "array"
          },
          "vulnerability_data": {
            "title": "Vulnerability Data",
            "type": "string"
          }
        },
        "required": [
          "total",
          "unique_packages",
          "production",
          "development",
          "other",
          "unconstrained",
          "external_sources",
          "dependencies",
          "dependencies_truncated",
          "manifests",
          "manifests_not_analyzed",
          "lockfiles",
          "ecosystems",
          "version_conflicts",
          "vulnerability_data",
          "notes"
        ],
        "title": "DependencyAnalysis",
        "type": "object"
      },
...
      "DocumentationAnalysis": {
        "properties": {
          "comment_density": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment Density"
          },
          "docs_directory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectoryInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "examples_directory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DirectoryInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/PresenceCheck"
            },
            "title": "Files",
            "type": "array"
          },
          "license_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License File"
          },
          "license_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Name"
          },
          "license_spdx": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Spdx"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "python_docstrings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocstringCoverage"
              },
              {
                "type": "null"
              }
            ]
          },
          "readme": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReadmeInfo"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "readme",
          "files",
          "license_spdx",
          "license_name",
          "license_file",
          "docs_directory",
          "examples_directory",
          "comment_density",
          "python_docstrings",
          "notes"
        ],
        "title": "DocumentationAnalysis",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HealthReport": {
        "properties": {
          "dimensions": {
            "items": {
              "$ref": "#/components/schemas/HealthDimension"
            },
            "title": "Dimensions",
            "type": "array"
          },
          "methodology": {
            "title": "Methodology",
            "type": "string"
          },
          "overall": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall"
          }
        },
        "required": [
          "overall",
          "dimensions",
          "methodology"
        ],
        "title": "HealthReport",
        "type": "object"
      },
...
      "KeyModule": {
        "properties": {
          "description": {
            "maxLength": 600,
            "title": "Description",
            "type": "string"
          },
          "path": {
            "maxLength": 300,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "description"
        ],
        "title": "KeyModule",
        "type": "object"
      },
      "LanguageAnalysis": {
        "properties": {
          "counted_lines": {
            "title": "Counted Lines",
            "type": "integer"
          },
          "estimated_lines": {
            "title": "Estimated Lines",
            "type": "integer"
          },
          "extensions": {
            "items": {
              "$ref": "#/components/schemas/ExtensionStat"
            },
            "title": "Extensions",
            "type": "array"
          },
          "github_breakdown": {
            "items": {
              "$ref": "#/components/schemas/GitHubLanguageShare"
            },
            "title": "Github Breakdown",
            "type": "array"
          },
          "languages": {
            "items": {
              "$ref": "#/components/schemas/LanguageStat"
            },
            "title": "Languages",
            "type": "array"
          },
          "methodology": {
            "title": "Methodology",
            "type": "string"
          },
          "primary_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Language"
          },
          "sampled_files": {
            "title": "Sampled Files",
            "type": "integer"
          },
          "total_lines": {
            "title": "Total Lines",
            "type": "integer"
          }
        },
        "required": [
          "primary_language",
          "github_breakdown",
          "languages",
          "extensions",
          "total_lines",
          "counted_lines",
          "estimated_lines",
          "sampled_files",
          "methodology"
        ],
        "title": "LanguageAnalysis",
        "type": "object"
      },
...
      "RepositoryMetadata": {
        "description": "Repository facts exactly as reported by the GitHub REST API.",
        "properties": {
          "archived": {
            "default": false,
            "title": "Archived",
            "type": "boolean"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "default_branch": {
            "title": "Default Branch",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "disabled": {
            "default": false,
            "title": "Disabled",
            "type": "boolean"
          },
          "forks": {
            "title": "Forks",
            "type": "integer"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "homepage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Homepage"
          },
          "html_url": {
            "title": "Html Url",
            "type": "string"
          },
          "is_fork": {
            "default": false,
            "title": "Is Fork",
            "type": "boolean"
          },
          "is_template": {
            "default": false,
            "title": "Is Template",
            "type": "boolean"
          },
          "license_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Name"
          },
          "license_spdx": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Spdx"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "open_issues": {
            "title": "Open Issues",
            "type": "integer"
          },
          "owner": {
            "title": "Owner",
            "type": "string"
          },
          "owner_avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Avatar Url"
          },
          "owner_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Type"
          },
          "primary_language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Language"
          },
          "pushed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pushed At"
          },
          "size_kb": {
            "title": "Size Kb",
            "type": "integer"
          },
          "stars": {
            "title": "Stars",
            "type": "integer"
          },
          "topics": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Topics",
            "type": "array"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "watchers": {
            "title": "Watchers",
            "type": "integer"
          }
        },
        "required": [
          "owner",
          "name",
          "full_name",
          "html_url",
          "default_branch",
          "stars",
          "forks",
          "watchers",
          "open_issues",
          "size_kb"
        ],
        "title": "RepositoryMetadata",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Deterministic static analysis of public GitHub repositories.",
    "title": "RepoLens API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/ai/summary": {
      "post": {
        "description": "Optional AI narrative built from the deterministic report (never from raw code).\n\nReturns 503 `ai_unavailable` when no provider is configured.",
        "operationId": "ai_summary_api_ai_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AISummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ai Summary"
      }
    },
    "/api/analyze": {
      "post": {
        "description": "Analyze a public repository and return the complete report.",
        "operationId": "analyze_api_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze"
      }
    },
    "/api/analyze/stream": {
      "post": {
        "description": "Same as /analyze, but streams progress events so clients can show the current stage.\n\nEach line is one JSON object:\n  {\"type\": \"progress\", \"stage\": \"tree\", \"message\": \"...\"}\n  {\"type\": \"result\", \"report\": {...}}\n  {\"type\": \"error\", \"error\": {\"code\": \"...\", \"message\": \"...\", \"details\": {...}}}",
        "operationId": "analyze_stream_api_analyze_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/x-ndjson": {}
            },
            "description": "Newline-delimited JSON events: progress, then result or error."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Stream"
      }
    },
...
    "/api/repositories/{owner}/{name}": {
      "get": {
        "operationId": "repository_metadata_api_repositories__owner___name__get",
        "parameters": [
          {
            "in": "path",
            "name": "owner",
            "required": true,
            "schema": {
              "title": "Owner",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepositoryMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Repository Metadata"
      }
    }
  }
}